requirements.txt
requests==2.32.3
tqdm>=4.66.5
pyinstaller>=6.3
pynsist>=2.8
